requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14
drf-yasg>=1.21
